drawing_validator/requirements.txt
pdfplumber
numpy
pandas
opencv-python
pymupdf
pytesseract
pillow
streamlit
fastapi
uvicorn[standard]
python-multipart